Add files via upload
In this notebook, I have performed some data analysis on the comments from tweets dataset that describes the way people talk about the stock market leveraging the natural language processing techniques to complete the task. Feel free to go through any piece and enjoy the rest of the work for projects or any other work.